Crate blazor_website

Blazor Website

CSCI 1260 website project for a C# Blazor Website

Uses leptos under the hood

Design

This website is a single-page-application utilizing several layers of fine-grained reactivity to re-render the entire body or just a singular character.

Modules

album

body

section

song

utils

Functions

Header

A simple header with typed text.

MainMenuPage

The page that gives the user all entry-level options. This includes viewing the list of albums and creating new albums

ReturnToMainMenu

A button that returns to the MainMenu page