Module blazor_website::album::album_page

Structs

ViewAlbumPageProps

Props for the ViewAlbumPage component.

Functions

ViewAlbumPage

A page/view that has an interactive album display with the normal typed text. There are options to add and remove songs to the playlist here.