Function blazor_website::album::album_page::ViewAlbumPage

pub fn ViewAlbumPage(props: ViewAlbumPageProps) -> impl IntoView

A page/view that has an interactive album display with the normal typed text. There are options to add and remove songs to the playlist here.

Arguements

playlist - ReadSignal for Playlist to view and edit

set_playlist - WriteSignal for Playlist to edit, should be the same as Playlist

set_page - WriteSignal for global page

Required Props

playlist: ReadSignal<Playlist>

set_playlist: WriteSignal<Playlist>

set_page: WriteSignal<Pages>