Struct blazor_website::album::album_page::ViewAlbumPageProps

pub struct ViewAlbumPageProps {
    pub playlist: ReadSignal<Playlist>,
    pub set_playlist: WriteSignal<Playlist>,
    pub set_page: WriteSignal<Pages>,
}

Props for the ViewAlbumPage component.

A page/view that has an interactive album display with the normal typed text. There are options to add and remove songs to the playlist here.

Arguements

playlist - ReadSignal for Playlist to view and edit

set_playlist - WriteSignal for Playlist to edit, should be the same as Playlist

set_page - WriteSignal for global page

Required Props

playlist: ReadSignal<Playlist>

set_playlist: WriteSignal<Playlist>

set_page: WriteSignal<Pages>

Fields
playlist: ReadSignal<Playlist>set_playlist: WriteSignal<Playlist>set_page: WriteSignal<Pages>
Implementations

impl ViewAlbumPageProps

pub fn builder() -> ViewAlbumPagePropsBuilder<((), (), ())>

Create a builder for building ViewAlbumPageProps. On the builder, call .playlist(...), .set_playlist(...), .set_page(...) to set the values of the fields. Finally, call .build() to create the instance of ViewAlbumPageProps.