Function blazor_website::utils::TypedText

pub fn TypedText(props: TypedTextProps) -> impl IntoView

Component that types out text supplied to it with a blinking cursor

Arguements

Required

text - Text to be typed

interval - Miliseconds between each letter

Optional

centered - (Def: false) Places a mock cursor to offset blinking cursor

stop - (Def: false) Stops the cursor after text has been rendered

delay - (Def: 0) Delay in milliseconds before typing

Panics

This function contains timeouts and intervals which have caused problems in the past when this function goes out of scope but the variables do not.

Notes

This function creates leptos_dom::helpers::set_timeout and leptos_dom::helpers::set_interval_with_handle. These call out to JS functions and don’t play nice with the safe rust side becuase both rust and JS are dealing with DOM values created in rust with leptos::create_signal

Required Props

text: String

interval: u64

Optional Props

centered: bool

stop: bool

delay: u64